+++ cypress/e2e/with_api/systems/systems.cy.ts
@@ -2,7 +2,7 @@ import { addCatalogueCategories } from '../catalogueCategories/functions';
 import { addCatalogueItem } from '../catalogueItems/functions';
 import { addItem } from '../items/functions';
 import { addManufacturer } from '../manufacturers/functions';
-import { addUsageStatuses } from '../usageStatus/functions';
+import { addUsageStatuses } from '../usageStatuses/functions';
 import {
   addSystems,
   copyToSystems,
